+++ Library/src/Multiplayer/Script/EntityScriptInterface.h
@@ -89,7 +89,7 @@ template <typename ScriptInterface, ComponentType Type> std::vector<ScriptInterf
         const auto& ComponentMap = *Entity->GetComponents();
         const auto ComponentKeys = ComponentMap.Keys();
 
-        for (int i = 0; i < ComponentKeys->Size(); ++i)
+        for (size_t i = 0; i < ComponentKeys->Size(); ++i)
         {
             ComponentBase* Component = ComponentMap[ComponentKeys->operator[](i)];
